requirements.txt
django>=1.11,<2
django-extensions
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-camel-case==1.1.0
PyPDF2
psycopg2-binary
requests